(PARTIAL! no FrameBufferMock.h) Merge pull request #3325 from npoltorapavlo/issue3321
Reorganize and drop unused external declarations. See issue #3321
