+++ .github/workflows/unit-tests.yml
@@ -143,6 +143,51 @@ jobs:
         with:
           path: rdkservices
 
+      - name: Generate external headers
+        # Empty headers to mute errors
+        run: >
+          cd "${{github.workspace}}/rdkservices/Tests/" &&
+          mkdir -p
+          headers
+          headers/audiocapturemgr
+          headers/rdk/ds
+          headers/rdk/iarmbus
+          headers/rdk/iarmmgrs-hal
+          headers/ccec/drivers &&
+          cd headers &&
+          touch
+          audiocapturemgr/audiocapturemgr_iarm.h
+          ccec/drivers/CecIARMBusMgr.h
+          rdk/ds/audioOutputPort.hpp
+          rdk/ds/compositeIn.hpp
+          rdk/ds/dsDisplay.h
+          rdk/ds/dsError.h
+          rdk/ds/dsMgr.h
+          rdk/ds/dsTypes.h
+          rdk/ds/dsUtl.h
+          rdk/ds/exception.hpp
+          rdk/ds/hdmiIn.hpp
+          rdk/ds/host.hpp
+          rdk/ds/list.hpp
+          rdk/ds/manager.hpp
+          rdk/ds/sleepMode.hpp
+          rdk/ds/videoDevice.hpp
+          rdk/ds/videoOutputPort.hpp
+          rdk/ds/videoOutputPortConfig.hpp
+          rdk/ds/videoOutputPortType.hpp
+          rdk/ds/videoResolution.hpp
+          rdk/iarmbus/libIARM.h
+          rdk/iarmbus/libIBus.h
+          rdk/iarmbus/libIBusDaemon.h
+          rdk/iarmmgrs-hal/deepSleepMgr.h
+          rdk/iarmmgrs-hal/mfrMgr.h
+          rdk/iarmmgrs-hal/pwrMgr.h
+          rdk/iarmmgrs-hal/sysMgr.h
+          framebuffer-api.h
+          libudev.h
+          rfcapi.h
+          telemetry_busmessage_sender.h
+
       - name: Configure rdkservices
         run: >
           cmake
@@ -150,11 +195,19 @@ jobs:
           -B build/rdkservices
           -DCMAKE_INSTALL_PREFIX="${{github.workspace}}/install/usr"
           -DCMAKE_MODULE_PATH="${{github.workspace}}/install/tools/cmake"
-          -DCMAKE_CXX_FLAGS="-I ${{github.workspace}}/rdkservices/Tests/headers
+          -DCMAKE_CXX_FLAGS="
+          -I ${{github.workspace}}/rdkservices/Tests/headers
           -I ${{github.workspace}}/rdkservices/Tests/headers/audiocapturemgr
           -I ${{github.workspace}}/rdkservices/Tests/headers/rdk/ds
           -I ${{github.workspace}}/rdkservices/Tests/headers/rdk/iarmbus
           -I ${{github.workspace}}/rdkservices/Tests/headers/rdk/iarmmgrs-hal
+          -I ${{github.workspace}}/rdkservices/Tests/headers/ccec/drivers
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/devicesettings.h
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/FrameBuffer.h
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/Iarm.h
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/Rfc.h
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/Telemetry.h
+          -include ${{github.workspace}}/rdkservices/Tests/mocks/Udev.h
           --coverage -Wall -Werror -Wno-unused-parameter -Wno-unused-result -Wl,-wrap,system -Wl,-wrap,popen
           -DENABLE_TELEMETRY_LOGGING
           -DUSE_IARMBUS
