--- Tests/headers/rdk/ds/list.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <vector>
-
-namespace device {
-
-template <class T>
-using List = std::vector<T>;
-
-}
