--- Tests/headers/rdk/iarmmgrs-hal/deepSleepMgr.h
@@ -1,31 +1,0 @@
-#pragma once
-
-#define IARM_BUS_DEEPSLEEPMGR_NAME "DEEPSLEEPMgr"
-
-#define IARM_BUS_DEEPSLEEPMGR_API_GetLastWakeupReason "GetLastWakeupReason"
-#define IARM_BUS_DEEPSLEEPMGR_API_GetLastWakeupKeyCode "GetLastWakeupKeycode"
-
-typedef enum _DeepSleep_WakeupReason_t {
-    DEEPSLEEP_WAKEUPREASON_IR = 0,
-    DEEPSLEEP_WAKEUPREASON_RCU_BT,
-    DEEPSLEEP_WAKEUPREASON_RCU_RF4CE,
-    DEEPSLEEP_WAKEUPREASON_GPIO,
-    DEEPSLEEP_WAKEUPREASON_LAN,
-    DEEPSLEEP_WAKEUPREASON_WLAN,
-    DEEPSLEEP_WAKEUPREASON_TIMER,
-    DEEPSLEEP_WAKEUPREASON_FRONT_PANEL,
-    DEEPSLEEP_WAKEUPREASON_WATCHDOG,
-    DEEPSLEEP_WAKEUPREASON_SOFTWARE_RESET,
-    DEEPSLEEP_WAKEUPREASON_THERMAL_RESET,
-    DEEPSLEEP_WAKEUPREASON_WARM_RESET,
-    DEEPSLEEP_WAKEUPREASON_COLDBOOT,
-    DEEPSLEEP_WAKEUPREASON_STR_AUTH_FAILURE,
-    DEEPSLEEP_WAKEUPREASON_CEC,
-    DEEPSLEEP_WAKEUPREASON_PRESENCE,
-    DEEPSLEEP_WAKEUPREASON_VOICE,
-    DEEPSLEEP_WAKEUPREASON_UNKNOWN
-} DeepSleep_WakeupReason_t;
-
-typedef struct _IARM_Bus_DeepSleepMgr_WakeupKeyCode_Param_t {
-    unsigned int keyCode;
-} IARM_Bus_DeepSleepMgr_WakeupKeyCode_Param_t;
